requirements.txt
dotenv
discord
requests
Pillow
python-dateutil
